add function calling LLM base class  (#12222)
