+++ llama-index-core/llama_index/core/agent/function_calling/step.py
@@ -246,7 +246,7 @@ class FunctionCallingAgentWorker(BaseAgentWorker):
             verbose=self._verbose,
             allow_parallel_tool_calls=self.allow_parallel_tool_calls,
         )
-        tool_calls = self._llm._get_tool_calls_from_response(
+        tool_calls = self._llm.get_tool_calls_from_response(
             response, error_on_no_tool_call=False
         )
         if not self.allow_parallel_tool_calls and len(tool_calls) > 1:
@@ -314,7 +314,7 @@ class FunctionCallingAgentWorker(BaseAgentWorker):
             verbose=self._verbose,
             allow_parallel_tool_calls=self.allow_parallel_tool_calls,
         )
-        tool_calls = self._llm._get_tool_calls_from_response(
+        tool_calls = self._llm.get_tool_calls_from_response(
             response, error_on_no_tool_call=False
         )
         if not self.allow_parallel_tool_calls and len(tool_calls) > 1:
